+++ src/app/file-manager/dbstate/appState.ts
@@ -1,0 +1,5 @@
+import { DownloadableState } from '../reducers/downloadables.reducer';
+
+export interface AppState {
+  downloadables: DownloadableState;
+}
